[MOVE/#115] 히스토리뷰 네이밍 수정
